Module rusty_secrets::dss::thss

Implements the ThSS threshold secret sharing scheme.

This scheme satisfies the following security properties:

Property 	Satisifed? 	Description

Basic 	Yes 	Basic correctness: If you attempt to recover a secret from an authorized set of shares that were obtained by sharing out a secret M using an access structure A, you're sure to get back A and M.
Note: in this implementation A is not actually returned, but definitely could.

Priv1 	Yes 	Standard privacy notation: When the coins are used by the dealer are uniformly random, unauthorized sets of shares have no computationally extractable information about the underlying secret.

Priv2 	No 	Privacy for deterministic or hedged schemes: extract whatever entropy one can from the underlying secret. If it’s adequate, no additional randomness is needed in order to achieve a meaningful notion of privacy.

Auth1 	No 	A share obtained from an honest dealer commits it to a single underlying secret: that and only that value can be recovered.

Auth2 	No 	A share obtained even from a dishonest dealer commits it to a single underlying secret: that and only that value might be recovered. Implies Auth1.

ErrDet 	Yes 	An inauthentic set of shares produced by an adversary will be flagged as such when fed to the recovery algorithm.

Repro 	No 	Share reproducible: The scheme can produce shares in a deterministic way.

Structs

MetaData
A share's public metadata.

Share
A share identified by an id, a threshold k, a number of total shares n, the data held in the share, and the share's metadata.

Functions

recover_secret
Recovers the secret from a k-out-of-n secret sharing scheme (ThSS).

split_secret
Performs threshold k-out-of-n secret sharing using the ThSS scheme.